fix: Scores: prettyPrintScoreboard renamed and moved to scores-graphics TU
- prettyPrintScoreboard() renamed to ScoreboardOverlay().
