+++ src/headers/scores-graphics.hpp
@@ -1,4 +1,12 @@
 #ifndef SCORESGRAPHICS_H
 #define SCORESGRAPHICS_H
 
+#include <string>
+
+namespace Scoreboard {
+namespace Graphics {
+std::string ScoreboardOverlay();
+}
+} // namespace Scoreboard
+
 #endif
